수정

+++ src/main/java/com/team/thebox/dao/MovieDAO.java
@@ -22,7 +22,7 @@ public interface MovieDAO {
 
     int insertMovieReply(MovieReply reply);
 
-    List<MovieReply> selectOneMovieReply(int movno);
+    List<MovieReply> selectOneMovieReply(Long movno);
 
     List<String> selectMovieTitle();
