+++ src/main/java/com/team/thebox/dao/admin/AMovieDAOImpl.java
@@ -72,7 +72,7 @@ public class AMovieDAOImpl implements AMovieDAO {
     }
 
     @Override
-    public List<MovieReply> selectOneMovieReply(int movno) {
+    public List<MovieReply> selectOneMovieReply(Long movno) {
         return movieReplyRepository.findByMovnoOrderByRegdateAsc(movno);
     }
